Added Repeat Message Filter
When enabled, the bot can now look for similar messages in the past minute of them being posted. If the message is similar to any of the previous ones sent by the user in the last minute, the bot will attempt to delete the new message.

Cleaned up code, fixed bug with bot reading its own messages.