Added Caps Filter and Permissions
The bot can now filter messages that are majority of capitalization. If the message is above 15 characters in length and 2/3 or greater of capitalized letters the bot will delete the message.

The bot will now look if the user using certain commands has permissions to. When setting up filters the user of the commands must have the appropriate permissions, whether that is either administrator or manage messages.

Cleaned up some code.